lunabot-nonebot/requirements.txt
aiohttp
aiosmtplib
aiosqlite
bs4
emoji==2.0.0
ffmpeg
matplotlib
mistune
nonebot2[fastapi]
nonebot-adapter-onebot
nonebot_plugin_htmlrender
nonebot_plugin_picstatus
nonebot_plugin_apscheduler
opencv-python-headless
orjson
pillow
pilmoji[requests]==2.0.4
pydub
python-dotenv
PyYAML
RapidFuzz
requests
sekaiworld.scores
selenium
setproctitle
tenacity
uvloop
zhconv
zhon
zstandard
colour-science